requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
chromadb
openai
langchain-community